+++ color.py
@@ -1,0 +1,8 @@
+class COLOR:
+    FIN = '\033[0m'   # hay que ponerlo al final de cada frase o si no el resto del programa usará el color anterior
+    ORADOR = '\033[93m' # orador
+    ALGOMALO = '\033[91m'     # algo malo en el gameplay
+    MONOLOGO = '\033[94m'     # monologo personaje
+    BLANCO = '\033[7m'      # fondo blanco
+    MORADO = '\033[45m'
+    WARNINGACCION = '\033[41m' # fondo rojo
