+++ cuadrado.py
@@ -1,0 +1,10 @@
+from forma import Forma
+
+class Cuadrado(Forma):
+    def __init__(self):
+        super().__init__()
+        self.norte = None
+        self.sur = None
+        self.este = None
+        self.oeste = None
+        self.orientaciones = []
