+++ forma.py
@@ -1,0 +1,41 @@
+import random
+from point import Point
+
+class Forma:
+    def __init__(self):
+        self.orientaciones = []
+        self.num = None
+        self.punto = Point(0,0)
+        self.extent = Point(0, 0)
+
+    def agregarOrientacion(self, orientacion):
+        self.orientaciones.append(orientacion)
+
+    def eliminarOrientacion(self, orientacion):
+        self.orientaciones.remove(orientacion)
+
+    def ponerElementoEnOrientacion(self, elemento, orientacion):
+        orientacion.poner(elemento, self)
+
+    def obtenerElementoEnOrientacion(self, orientacion):
+        return orientacion.obtenerElemento(self)
+
+    def recorrer(self, func):
+        for orientacion in self.orientaciones:
+            orientacion.recorrer(func, self)
+
+    def calcularPosicion(self):
+        for orientacion in self.orientaciones:
+            orientacion.calcularPosicionDesde(self)
+
+    def caminarAleatorio(self, ente):
+        orientacion = self.obtenerOrientacionAleatoria()
+        print(f"Orientacion aleatoria: {orientacion}")
+        orientacion.caminarAleatorio(ente, self)
+
+    def obtenerOrientacionAleatoria(self):
+        return random.choice(self.orientaciones)
+
+    def aceptar(self, unVisitor):
+        for orientacion in self.orientaciones:
+            orientacion.aceptar(unVisitor, self)
