+++ hoja.py
@@ -1,0 +1,6 @@
+from elementoMapa import ElementoMapa
+
+# LEAF
+class Hoja(ElementoMapa):
+    def __init__(self):
+        super().__init__()
