+++ lab4hab.json
@@ -1,0 +1,19 @@
+{
+	"forma":"cuadrado",
+    "laberinto":[
+	{"tipo":"habitacion", "num":1, "hijos":[]},
+	{"tipo":"habitacion", "num":2, "hijos":[]},
+	{"tipo":"habitacion", "num":3, "hijos":[]},
+	{"tipo":"habitacion", "num":4, "hijos":[]}
+	],
+	"puertas":[
+	[1, "Este", 2, "Oeste"], [2,"Sur", 4, "Norte"],
+	[3, "Este", 4, "Oeste"], [3, "Norte", 1, "Sur"]
+	],
+	"bichos":[
+		{"modo":"Agresivo", "posicion":1},
+		{"modo":"Agresivo", "posicion":3},
+		{"modo":"Perezoso", "posicion":2},
+		{"modo":"Agresivo", "posicion":4}
+	]
+}
