+++ laberintoBuilder.py
@@ -1,11 +1,23 @@
+from puerta import Puerta
+from norte import Norte
+from sur import Sur
+from este import Este
+from oeste import Oeste
+from habitacion import Habitacion
+from juego import Juego
+from agresivo import Agresivo
+from perezoso import Perezoso
+from tunel import Tunel
+from laberinto import Laberinto
+from cuadrado import Cuadrado
+from pared import Pared
+from ente import Personaje
+from bicho import Bicho
+
 import copy
-import time
-
-#from Juego import Laberinto, Habitacion, Pared, Puerta, Cuadrado, Norte, Sur, Este, Oeste, Bicho, Agresivo, Perezoso, Tunel, Juego
-from Juego import *
 
 
-class LabBuilder:
+class LaberintoBuilder:
     def __init__(self):
         self.laberinto = None
         self.juego = None
@@ -13,7 +25,7 @@ class LabBuilder:
     def fabricarJuego(self):
         self.juego = Juego()
         self.juego.prototipo = self.laberinto
-        self.juego.laberinto = copy.deepcopy(self.juego.laberinto)
+        self.juego.laberinto = copy.deepcopy(self.juego.prototipo)
 
     def fabricarLaberinto(self):
         self.laberinto = Laberinto()
@@ -22,25 +34,26 @@ class LabBuilder:
         hab = Habitacion(num)
         hab.forma = self.fabricarForma()
         hab.forma.num = num
+        # hab.agregarOrientacion(self.fabricarNorte())
+        # hab.agregarOrientacion(self.fabricarSur())
+        # hab.agregarOrientacion(self.fabricarEste())
+        # hab.agregarOrientacion(self.fabricarOeste())
 
-        for h in hab.forma.orientaciones:
-            hab.ponerElementoEnOrientacion(self.fabricarPared(), h)
-
+        for i in hab.forma.orientaciones:
+            hab.ponerElementoEnOrientacion(self.fabricarPared(), i)
         self.laberinto.agregarHabitacion(hab)
-
         return hab
-
 
     def fabricarPared(self):
         return Pared()
 
-    def fabricarPuerta(self, lado1, ob1, lado2, ob2):
+    def fabricarPuerta(self, lado1, obj1, lado2, obj2):
         hab1 = self.laberinto.obtenerHabitacion(lado1)
         hab2 = self.laberinto.obtenerHabitacion(lado2)
         puerta = Puerta(hab1, hab2)
 
-        objOr1 = self.obtenerObjeto(ob1)
-        objOr2 = self.obtenerObjeto(ob2)
+        objOr1 = self.obtenerObjeto(obj1)
+        objOr2 = self.obtenerObjeto(obj2)
         hab1.ponerElementoEnOrientacion(puerta, objOr1)
         hab2.ponerElementoEnOrientacion(puerta, objOr2)
 
@@ -63,6 +76,7 @@ class LabBuilder:
         forma.agregarOrientacion(self.fabricarSur())
         forma.agregarOrientacion(self.fabricarEste())
         forma.agregarOrientacion(self.fabricarOeste())
+
         return forma
 
     def fabricarNorte(self):
@@ -92,17 +106,15 @@ class LabBuilder:
     def obtenerJuego(self):
         return self.juego
 
+    def fabricarBicho(self, modo, posicion):
+        if modo=='Agresivo':
+            bicho = self.fabricarBichoAgresivo()
+        if modo=='Perezoso':
+            bicho = self.fabricarBichoPerezoso()
+        hab = self.laberinto.obtenerHabitacion(posicion)
+        hab.entrar(bicho)
+        self.juego.agregarBicho(bicho)
+
     def fabricarTunelEn(self, unCont):
         tunel = Tunel(None)
         unCont.agregarHijo(tunel)
-
-    def fabricarBicho(self, modo, posicion):
-        if modo == 'Agresivo':
-            bicho = self.fabricarBichoAgresivo()
-        if modo == 'Perezoso':
-            bicho = self.fabricarBichoPerezoso()
-
-        hab = self.laberinto.obtenerHabitacion(posicion)
-
-        hab.entrar(bicho)
-        self.juego.agregarBicho(bicho)
