+++ modo.py
@@ -1,0 +1,18 @@
+# Modo de Bichos (Strategy)
+class Modo:
+    def __init__(self):
+        pass
+
+    def actuar(self, bicho):
+        self.dormir(bicho)
+        self.caminar(bicho)
+        self.atacar(bicho)
+
+    def dormir(self, bicho):
+        pass
+
+    def caminar(self, bicho):
+        pass
+
+    def atacar(self, bicho):
+        pass
