+++ oeste.py
@@ -1,0 +1,32 @@
+from orientacion import Orientacion
+from point import Point
+
+class Oeste(Orientacion):
+    _instance = None
+
+    def __new__(cls):
+        if cls._instance is None:
+            cls._instance = super().__new__(cls)
+        return cls._instance
+
+    def poner(self, elemento, contenedor):
+        contenedor.oeste = elemento
+
+    def recorrer(self, func, contenedor):
+        if contenedor.oeste is not None:
+            func(contenedor.oeste)
+
+    def obtenerElemento(self, forma):
+        return forma.oeste
+
+    def caminarAleatorio(self, ente, forma):
+        forma.oeste.entrar(ente)
+
+    def aceptar(self, unVisitor, forma):
+        forma.oeste.aceptar(unVisitor)
+    def calcularPosicionDesde(self, forma):
+        unPunto = Point(forma.punto.x-1,forma.punto.y)
+        forma.oeste.calcularPosicionDesdeEn(forma,unPunto)
+
+    def __repr__(self):
+        return "oeste"
