+++ point.py
@@ -1,0 +1,4 @@
+class Point:
+    def __init__(self, x_init, y_init):
+        self.x = x_init
+        self.y = y_init
