+++ visitor.py
@@ -1,0 +1,15 @@
+class Visitor:
+    def visitarHabitacion(self, habitacion):
+        pass
+
+    def visitarPared(self, pared):
+        pass
+
+    def visitarPuerta(self, puerta):
+        pass
+
+    def visitarBomba(self, bomba):
+        pass
+
+    def visitarTunel(self, tunel):
+        pass
